fix: [PL-42886]: Revert the accountRingInfo metric (#55706)
* fix: [PL-42886]: Revert "fix: [PL-42592]: Modified the time value to be constant for an iteration in AccountRingInfo iterator (#55586) (#55608)"

This reverts commit 28e262ea44848c7c3cad1567a0f5e0ba41a04f1a.

* fix: [PL-42886]: Revert "fix: [PL-42502]: Added a new metric ACCOUNT_RING_INFO to get Account-ring information  (#55389)"

This reverts commit 1e4d3e91563d534292261ff4f1f0091bc4eedc41.

* fix: [PL-42886]:Set the correct build number

* fix: [PL-42886]: Set the delegate version back to current

* fix: [PL-42886]: Add back the delegateExpiryIterator config
