+++ 360-cg-manager/iterator_config.json
@@ -378,15 +378,5 @@
     "targetIntervalInSeconds": 86400,
     "throttleIntervalInSeconds": 0,
     "iteratorMode": "REDIS_BATCH"
-  },
-  {
-    "name": "AccountRingInformation",
-    "enabled": true,
-    "threadPoolSize": 1,
-    "threadPoolIntervalInSeconds": 30,
-    "nextIterationMode": "TARGET",
-    "targetIntervalInSeconds": 7200,
-    "throttleIntervalInSeconds": 0,
-    "iteratorMode": "REDIS_BATCH"
   }
 ]
