--- 400-rest/src/main/java/io/harness/delegate/heartbeat/AccountRingInfoIterator.java
@@ -1,96 +1,0 @@
-/*
- * Copyright 2023 Harness Inc. All rights reserved.
- * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
- * that can be found in the licenses directory at the root of this repository, also available at
- * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
- */
-
-package io.harness.delegate.heartbeat;
-
-import static io.harness.mongo.iterator.MongoPersistenceIterator.SchedulingType.REGULAR;
-
-import static java.time.Duration.ofMinutes;
-
-import io.harness.iterator.IteratorExecutionHandler;
-import io.harness.iterator.IteratorPumpAndRedisModeHandler;
-import io.harness.iterator.PersistenceIteratorFactory;
-import io.harness.mongo.iterator.MongoPersistenceIterator;
-import io.harness.mongo.iterator.filter.MorphiaFilterExpander;
-import io.harness.mongo.iterator.provider.MorphiaPersistenceProvider;
-
-import software.wings.beans.Account;
-import software.wings.beans.Account.AccountKeys;
-
-import com.google.inject.Inject;
-import com.google.inject.Singleton;
-import java.time.Duration;
-import java.util.concurrent.TimeUnit;
-import lombok.extern.slf4j.Slf4j;
-
-@Singleton
-@Slf4j
-public class AccountRingInfoIterator
-    extends IteratorPumpAndRedisModeHandler implements MongoPersistenceIterator.Handler<Account> {
-  private static final Duration ACCEPTABLE_NO_ALERT_DELAY = ofMinutes(10);
-  @Inject private io.harness.iterator.PersistenceIteratorFactory persistenceIteratorFactory;
-  @Inject private MorphiaPersistenceProvider<Account> persistenceProvider;
-  @Inject private AccountRingInfoMetricHelper accountRingInfoMetricHelper;
-  private static Long time;
-
-  @Override
-  protected void createAndStartIterator(
-      PersistenceIteratorFactory.PumpExecutorOptions executorOptions, Duration targetInterval) {
-    iterator = (MongoPersistenceIterator<Account, MorphiaFilterExpander<Account>>)
-                   persistenceIteratorFactory.createPumpIteratorWithDedicatedThreadPool(executorOptions, Account.class,
-                       MongoPersistenceIterator.<Account, MorphiaFilterExpander<Account>>builder()
-                           .clazz(Account.class)
-                           .fieldName(AccountKeys.accountRingInfoIteration)
-                           .filterExpander(getFilterQuery())
-                           .targetInterval(targetInterval)
-                           .acceptableNoAlertDelay(ACCEPTABLE_NO_ALERT_DELAY)
-                           .handler(this)
-                           .schedulingType(REGULAR)
-                           .persistenceProvider(persistenceProvider)
-                           .redistribute(true));
-  }
-
-  @Override
-  protected void createAndStartRedisBatchIterator(
-      PersistenceIteratorFactory.RedisBatchExecutorOptions executorOptions, Duration targetInterval) {
-    iterator =
-        (MongoPersistenceIterator<Account, MorphiaFilterExpander<Account>>)
-            persistenceIteratorFactory.createRedisBatchIteratorWithDedicatedThreadPool(executorOptions, Account.class,
-                MongoPersistenceIterator.<Account, MorphiaFilterExpander<Account>>builder()
-                    .clazz(Account.class)
-                    .fieldName(AccountKeys.accountRingInfoIteration)
-                    .filterExpander(getFilterQuery())
-                    .targetInterval(targetInterval)
-                    .acceptableNoAlertDelay(ACCEPTABLE_NO_ALERT_DELAY)
-                    .handler(this)
-                    .schedulingType(REGULAR)
-                    .persistenceProvider(persistenceProvider));
-  }
-
-  @Override
-  public void registerIterator(IteratorExecutionHandler iteratorExecutionHandler) {
-    iteratorName = "AccountRingInformation";
-    iteratorExecutionHandler.registerIteratorHandler(iteratorName, this);
-  }
-
-  @Override
-  public void handle(Account account) {
-    long currentTime = System.currentTimeMillis();
-    if (time == null || TimeUnit.MILLISECONDS.toMinutes(currentTime - time) >= 30) {
-      time = currentTime;
-    }
-    log.debug("Account being handled {}", account.getAccountName());
-    accountRingInfoMetricHelper.addAccountRingInfoMetric(
-        account.getUuid(), account.getAccountName(), account.getRingName(), time);
-  }
-
-  private MorphiaFilterExpander<Account> getFilterQuery() {
-    return query -> {
-      query.and(query.criteria(AccountKeys.ringName).exists());
-    };
-  }
-}
