+++ 400-rest/src/main/java/io/harness/delegate/heartbeat/DelegateHeartBeatMetricsHelper.java
@@ -16,7 +16,10 @@ import static io.harness.metrics.impl.DelegateMetricsServiceImpl.HEARTBEAT_RESTA
 import static io.harness.metrics.impl.DelegateMetricsServiceImpl.HEARTBEAT_UNREGISTER_EVENT;
 
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.delegate.beans.DelegateRing;
+import io.harness.delegate.beans.DelegateRing.DelegateRingKeys;
 import io.harness.metrics.intfc.DelegateMetricsService;
+import io.harness.persistence.HPersistence;
 
 import software.wings.beans.Account;
 import software.wings.dl.GenericDbCache;
@@ -25,7 +28,6 @@ import software.wings.service.intfc.AccountService;
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.google.inject.Inject;
-import java.time.Clock;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -36,13 +38,15 @@ import lombok.extern.slf4j.Slf4j;
 @OwnedBy(PL)
 public class DelegateHeartBeatMetricsHelper {
   @Inject private AccountService accountService;
+  @Inject private HPersistence persistence;
   @Inject private GenericDbCache dbCache;
   @Inject private DelegateMetricsService delegateMetricsService;
-  @Inject protected Clock clock;
 
+  private final Cache<String, DelegateRing> delegateRingCache =
+      Caffeine.newBuilder().maximumSize(10).expireAfterWrite(30, TimeUnit.MINUTES).build();
   private final Cache<String, Long> lastHeartBeatTimeCache =
       Caffeine.newBuilder().maximumSize(5000L).expireAfterWrite(10, TimeUnit.MINUTES).build();
-  private final Long HEARTBEAT_RECORD_FREQUENCY = 20L;
+  private final Long HEARTBEAT_RECORD_FREQUENCY = 10L;
 
   private List<String> getAccountDetails(String accountId) {
     List<String> accountInfo = new ArrayList<>();
@@ -59,6 +63,17 @@ public class DelegateHeartBeatMetricsHelper {
     accountInfo.add(accountFromDB.getCompanyName());
     accountInfo.add(accountFromDB.getRingName());
     return accountInfo;
+  }
+
+  private DelegateRing getDelegateRing(String ringName) {
+    DelegateRing delegateRing = delegateRingCache.getIfPresent(ringName);
+    if (delegateRing != null) {
+      return delegateRing;
+    }
+    DelegateRing ringFromDB =
+        persistence.createQuery(DelegateRing.class).filter(DelegateRingKeys.ringName, ringName).get();
+    delegateRingCache.put(ringName, ringFromDB);
+    return ringFromDB;
   }
 
   public void addDelegateHeartBeatMetric(long time, String accountId, String orgId, String projectId,
@@ -80,9 +95,10 @@ public class DelegateHeartBeatMetricsHelper {
     }
     try {
       List<String> accountInfo = getAccountDetails(accountId);
+      DelegateRing delegateRing = (accountInfo.get(2) != null) ? getDelegateRing(accountInfo.get(2)) : null;
       delegateMetricsService.recordDelegateHeartBeatMetricsPerAccount(accountId, accountInfo.get(0), accountInfo.get(1),
-          orgId, projectId, delegateName, delegateId, delegateVersion, delegateConnectionStatus, delegateEventType,
-          isNg, isImmutable, lastHB, metricName);
+          delegateRing, orgId, projectId, delegateName, delegateId, delegateVersion, delegateConnectionStatus,
+          delegateEventType, isNg, isImmutable, lastHB, metricName);
       if (Objects.equals(delegateEventType, HEARTBEAT_DELETE_EVENT)
           || Objects.equals(delegateEventType, HEARTBEAT_UNREGISTER_EVENT)
           || Objects.equals(delegateEventType, HEARTBEAT_RESTART_EVENT)) {
