+++ 400-rest/src/main/java/software/wings/beans/Account.java
@@ -208,7 +208,6 @@ public class Account extends Base implements PersistentRegularIterable, NGMigrat
   @FdIndex private long delegateTelemetryPublisherIteration;
   @FdIndex private long delegateTaskRebroadcastIteration;
   @FdIndex private Long perpetualTaskRebalanceIteration;
-  @FdIndex private Long accountRingInfoIteration;
 
   // adding this to avoid kryo exception. Its not used anymore, check DEL-5047
   @Deprecated private long delegateTaskFailIteration;
@@ -548,11 +547,6 @@ public class Account extends Base implements PersistentRegularIterable, NGMigrat
       return;
     }
 
-    else if (AccountKeys.accountRingInfoIteration.equals(fieldName)) {
-      this.accountRingInfoIteration = nextIteration;
-      return;
-    }
-
     throw new IllegalArgumentException("Invalid fieldName " + fieldName);
   }
 
@@ -600,10 +594,6 @@ public class Account extends Base implements PersistentRegularIterable, NGMigrat
 
     else if (AccountKeys.perpetualTaskRebalanceIteration.equals(fieldName)) {
       return this.perpetualTaskRebalanceIteration;
-    }
-
-    else if (AccountKeys.accountRingInfoIteration.equals(fieldName)) {
-      return this.accountRingInfoIteration;
     }
 
     throw new IllegalArgumentException("Invalid fieldName " + fieldName);
