--- 400-rest/src/main/resources/metrics/metricDefinitions/account_ring_info_metrics.yaml
@@ -1,8 +1,0 @@
-name: Account Ring Information Metrics
-identifier: account_ring_info_metrics
-metricGroup: account_ring_info_entry
-metrics:
-  - metricName: account_ring_info
-    metricDefinition: Account Ring info
-    type: LastValue
-    unit: count
