--- 400-rest/src/main/resources/metrics/metricGroups/account_ring_info_metrics_group.yaml
@@ -1,12 +1,0 @@
-name: AccountRingInfo
-identifier: account_ring_info_entry
-labels:
-  - accountId
-  - accountName
-  - ringName
-  - delegateImageTag
-  - upgraderImageTag
-  - watcherVersion
-  - watcherJREVersion
-  - delegateJREVersion
-  - legacyDelegateVersion
