+++ 400-rest/src/main/resources/metrics/metricGroups/heartbeat_metrics_group.yaml
@@ -4,6 +4,12 @@ labels:
   - accountId
   - accountName
   - companyName
+  - ringName
+  - delegateImageTag
+  - upgraderImageTag
+  - watcherVersion
+  - watcherJREVersion
+  - delegateJREVersion
   - orgId
   - projectId
   - delegateName
