+++ 420-delegate-service/src/main/java/io/harness/metrics/beans/HeartbeatMetricContext.java
@@ -10,12 +10,20 @@ package io.harness.metrics.beans;
 import io.harness.metrics.AutoMetricContext;
 
 public class HeartbeatMetricContext extends AutoMetricContext {
-  public HeartbeatMetricContext(String accountId, String accountName, String companyName, String orgId,
-      String projectId, String delegateName, String delegateId, String delegateVersion, String delegateConnectionStatus,
-      String delegateEventType, boolean isNg, boolean isImmutable) {
+  public HeartbeatMetricContext(String accountId, String accountName, String companyName, String ringName,
+      String delegateImageTag, String upgraderImageTag, String watcherVersion, String watcherJREVersion,
+      String delegateJREVersion, String orgId, String projectId, String delegateName, String delegateId,
+      String delegateVersion, String delegateConnectionStatus, String delegateEventType, boolean isNg,
+      boolean isImmutable) {
     put("accountId", accountId);
     put("accountName", accountName);
     put("companyName", companyName);
+    put("ringName", ringName);
+    put("delegateImageTag", delegateImageTag);
+    put("upgraderImageTag", upgraderImageTag);
+    put("watcherVersion", watcherVersion);
+    put("watcherJREVersion", watcherJREVersion);
+    put("delegateJREVersion", delegateJREVersion);
     put("orgId", orgId);
     put("projectId", projectId);
     put("delegateName", delegateName);
